Set timestamp for column even if invalid
If the first column in a sweep is skipped due to being marked as
invalid, then the time stamp is left at the default value of zero. The
timestamp for the first column is subsequently used as the timestamp
for the entire sweep when producing ROS messages, so leaving this as
zero can produce a bogus ROS header.